requirements.txt
alive_progress==3.2.0
hydra-core==1.3.2
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
scikit_learn==1.6.1
scipy==1.15.2
torch==2.6.0
torch_summary==1.4.5
torchmetrics==1.6.1
torchview==0.2.6
torchvision==0.21.0
